backend/requirements.txt
celery==5.1.2
fastapi==0.68.1
httpx==0.19.0
psycopg2-binary==2.9.1
python-dotenv==0.19.1
python-jose==3.3.0
pyjwt==2.3.0
redis==3.5.3
sqlalchemy==1.4.25
uvicorn==0.15.0
